Function Index

Stability

The entries written on this page are automatically generated and DO NOT represent the currently supported API surface. PropCheck.jl is currently in a 0.x release; feel free to use anything you can find here, but know that (for now) only symbols that are exported are expected to stick around (they too may change, but I don't expect the underlying functionality to vanish entirely).

Nevertheless, I don't really expect things to change too much - the package is already complicated as is.

PropCheck.AbstractIntegrated

PropCheck.ExtentIntegrated

PropCheck.Flatten

PropCheck.Integrated

PropCheck.IntegratedConst

PropCheck.IntegratedRange

PropCheck.IntegratedUnique

PropCheck.IntegratedVal

PropCheck.Tree

PropCheck.UniqueIterator

Base.Iterators.map

Base.filter

Base.filter

Base.map

PropCheck.generate

PropCheck.iconst

PropCheck.isample

PropCheck.isample

PropCheck.itype

PropCheck.iunique

PropCheck.ival

PropCheck.noshrink

PropCheck.root

PropCheck.shrink

PropCheck.shrinkTowards

PropCheck.str

PropCheck.subtrees

PropCheck.tuple

PropCheck.unfold

Function reference

Base.Iterators.map — Method

map(f, ::Tree)

Maps the function f over the nodes of the given Tree, returning a new true.

The subtrees of the tree returned by map are guaranteed to be unique.

Base.filter — Method

filter(p, i::AbstractIntegrated[, trim=false])

Filters i lazily such that all elements contained fulfill the predicate p, i.e. all elements for which p is false are removed.

trim controls whether subtrees are removed completely when the root doesn't fulfill the predicate or whether only that root should be skipped, still trying to shrink its subtrees. This trades performance (less shrinks to check) for quality (fewer/less diverse shrink values tried).

Base.filter — Method

filter(pred, ::Tree, trim=false)

Filters the given Tree by applying the predicate pred to each root.

Filtering is done lazily for subtrees.

trim controls whether subtrees that don't match the predicate should be pruned entirely, or only their roots should be removed from the resulting Tree. If trim is true, subtrees of subtrees that don't match the predicate are moved to the parent.

Base.map — Method

map(f, i::Integrated)

Maps f lazily over all elements in i, producing a new tree.

PropCheck.iconst — Method

iconst(x) -> AbstractIntegrated

A convenience constructor for creating an integrated shrinker.

Trees created by this do not shrink, and generate on the returned AbstractIntegrated will always produce x.

PropCheck.isample — Function

isample(x::AbstractRange[, shrink=shrinkTowards(first(x))]) -> AbstractIntegrated

A convenience constructor for creating an integrated shrinker.

Trees created by this shrink towards the first element of the range by default.

PropCheck.isample — Function

isample(x[, shrink=shrink]) -> AbstractIntegrated

A convenience constructor for creating an integrated shrinker.

Trees created by this shrink according to shrink, and generate on the returned Integrated will always produce an element of the collection x.

x needs to be indexable.

PropCheck.iunique — Method

iunique(x::Vector, shrink=shrink) -> AbstractIntegrated

A convenience constructor for creating an integrated shrinker.

This shrinker produces all unique values of x before producing a value it has produced before. The produced trees shrink according to shrink.

PropCheck.ival — Method

ival(x::T[, shrink=shrink]) -> AbstractIntegrated

A convenience constructor for creating integrated shrinkers, generating their values from a starting value.

Trees created by this function will always initially have x at their root, as well as shrink according to shrink.

PropCheck.noshrink — Method

noshrink(_::T) -> T[]

A shrinker that doesn't shrink its arguments.

PropCheck.root — Method

root(t::Tree)

Returns the singular root of the given Tree.

See also subtrees.

PropCheck.shrinkTowards — Function

shrinkTowards(to::T) -> (x::T -> T[...])

Constructs a shrinker function that shrinks given values towards to.

PropCheck.str — Function

str(len::ExtentIntegrated[, alphabet::AbstractIntegrated])

Generates a string using the given genLen as a generator for the length. The default alphabet is typemin(Char):""[1], which is all representable Char values.

PropCheck.subtrees — Method

subtrees(t::Tree)

Returns the (potentially lazy) subtrees of the given Tree.

See also root.

PropCheck.tuple — Method

tuple(genLen::AbstractIntegrated, genA::AbstractIntegrated)

Generates a tuple of a generated length, using the elements produced by genA.

PropCheck.unfold — Method

unfold(f, root)

Unfolds root into a Tree, by applying f to each root to create subtrees.

root is the new root of the returned Tree. f must return an iterable object with eltype egal to typeof(root).

PropCheck.Flatten — Type

Flatten with inferrable eltype. Requires the given iterators to have the same eltype.

PropCheck.Integrated — Type

Integrated{T} <: AbstractIntegrated{T}

A naive integrated shrinker, only providing extremely basic functionality for generating and shrinking Trees. Default fallback if no other, more specialized type exists.

Pending Redesign

This type is likely going to be redesigned in the future. Methods constructing it are not stable, and may be removed at any point.

PropCheck.IntegratedConst — Type

IntegratedConst{T,R,G} <: ExtentIntegrated{T}

An integrated shrinker describing a constant. The shrinker will always produce that value, which doesn't shrink.

PropCheck.IntegratedRange — Type

IntegratedRange{T,R,G,F} <: ExtentIntegrated{T}

An integrated shrinker describing a range of values.

The values created by this shrinker shrink according to the given shrinking function. The shrinking function must ensure that the produce values are always contained within the bounds of the given range.

PropCheck.IntegratedUnique — Type

IntegratedUnique(vec::Vector{ElT}, shrink::S) where {ElT,S}
IntegratedUnique{T,ElT,S} <: AbstractIntegrated{T}

An integrated shrinker, taking a vector vec. The shrinker will produce all unique values of vec in a random order before producing a value it returned before. The values produced by this shrinker shrink according to shrink.

PropCheck.IntegratedVal — Type

IntegratedVal(val::V, shrink::S) where {V,S}
IntegratedVal{T,V,S} <: AbstractIntegrated{T}

An integrated shrinker, taking a value val. The shrinker will always produce val, which shrinks according to shrink.

If V <: Number, shrinking functions given to this must produce values in

[typemin(V), v] if v > zero(V)

[v, typemax(V)] if v < zero(V)

no values if iszero(v)

This shrinker supports extent out of the box if V <: Number. For other types, you need to define extent(::IntegratedVal{Tree{T}})

PropCheck.Tree — Type

Tree{T}

A tree of T objects. The tree is inherently lazy; subtrees of subtrees are only generated on demand.

The subtrees field is intentionally untyped, to allow for various kinds of lazy subtree representations.

PropCheck.UniqueIterator — Type

UniqueIterator(itr, by=identity)

A lazy iterator over unique elements that have not been produced before.