Function Index

Stability

The entries written on this page are automatically generated and DO NOT represent the currently supported API surface. PropCheck.jl is currently in a 0.x release; feel free to use anything you can find here, but know that (for now) only symbols that are exported are expected to stick around (they too may change, but I don't expect the underlying functionality to vanish entirely).

Nevertheless, I don't really expect things to change too much - the package is already complicated as is.

PropCheck.ExtentIntegrated

PropCheck.Flatten

PropCheck.IntegratedConst

PropCheck.IntegratedRange

PropCheck.IntegratedUnique

PropCheck.IntegratedVal

PropCheck.UniqueIterator

Base.filter

Base.map

PropCheck.generate

PropCheck.iconst

PropCheck.isample

PropCheck.isample

PropCheck.itype

PropCheck.iunique

PropCheck.ival

PropCheck.noshrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrink

PropCheck.shrinkTowards

PropCheck.str

PropCheck.tuple

PropCheck.unfold

Function reference

Base.filter — Method

filter(p, i::Integrated[, trim=false])

Filters i lazily such that all elements contained fulfill the predicate p, i.e. all elements for which p is false are removed.

trim controls whether subtrees are removed completely when the root doesn't fulfill the predicate or whether only that root should be skipped, still trying to shrink its subtrees. This trades performance (less shrinks to check) for quality (fewer/less diverse shrink values tried).

Base.map — Method

map(f, i::Integrated)

Maps f lazily over all elements in i, producing a new tree.

PropCheck.generate — Function

generate(rng::AbstractRNG, ::Type{T}) where T -> T

Function to generate a single value of type T. Falls back to constructor inspection, which will generate values for ::Any typed arguments.

Types that have rand defined for them should forward to it here. Assumed to return an object of type T.

PropCheck.iconst — Method

iconst(x)

A convenience constructor for creating an integrated shrinker.

Trees created by this do not shrink, and generate on the returned AbstractIntegrated will always produce x.

PropCheck.isample — Function

isample(x::AbstractRange[, shrink=shrinkTowards(first(x))])

A convenience constructor for creating an integrated shrinker.

Trees created by this shrink towards the first element of the range by default.

PropCheck.isample — Function

isample(x[, shrink=shrink])

A convenience constructor for creating an integrated shrinker.

Trees created by this shrink according to shrink, and generate on the returned Integrated will always produce an element of the collection x.

x needs to be indexable.

PropCheck.itype — Method

itype(T::Type[, shrink=shrink])

A convenience constructor for creating integrated shrinkers, generating their values from a type.

Trees created by this function will have their elements shrink according to shrink.

PropCheck.iunique — Method

iunique(x::Vector, shrink=shrink)

A convenience constructor for creating an integrated shrinker.

This shrinker produces all unique values of x before producing a value it has produced before. The produced trees shrink according to shrink.

PropCheck.ival — Method

ival(x::T[, shrink=shrink])

A convenience constructor for creating integrated shrinkers, generating their values from a starting value.

Trees created by this function will always initially have x at their root, as well as shrink according to shrink.

PropCheck.noshrink — Method

noshrink(_::T) -> T[]

A shrinker that doesn't shrink its arguments.

PropCheck.shrink — Function

shrink(val::T) where T

Function to be used during shrinking of val. Must return an iterable of shrunk values, which can be lazy. If the returned iterable is empty, it's taken as a signal that the given value cannot shrink further.

Must never return a previously input value, i.e. no value val used as input should ever be produced by shrink(val) or subsequent applications of shrink on the produced elements. This will lead to infinite looping during shrinking.

PropCheck.shrink — Method

shrink(::Bool)

true shrinks to false. false produces no shrunk values.

PropCheck.shrink — Method

shrink(::Char)

Shrinks a character by shrinking its unicode codepoint.

Shrinks towards '\0'. '\0' produces no shrunk values.

PropCheck.shrink — Method

shrink(::String)

Shrinks a string by either dropping a character, or shrinking a character.

Shrinks towards the empty string, "". The empty string produces no shrunk values.

PropCheck.shrink — Method

shrink(::Tuple)

Shrinks a tuple by shrinking an element. The resulting tuples are always of the same length as the input tuple.

PropCheck.shrink — Method

shrink(::AbstractDict)

Shrinks a dictionary by shrinking keys, values and by dropping an entry. The empty dictionary doesn't shrink.

PropCheck.shrink — Method

shrink(::T) where T <: AbstractFloat

Shrinks an AbstractFloat.

Shrinks towards iszero(T). Inf, NaN and zero(T) produce no shrunk values. Positive numbers produce their negative counterparts.

PropCheck.shrink — Method

shrink(::AbstractRange)

Shrinks a range by splitting it in two in the middle. Ranges containing only one element or less don't shrink.

PropCheck.shrink — Method

shrink(::T) where T <: Signed

Shrinks a signed value by shrinking its absolute value like an unsigned and then also negating that. Both the absolute and the negated value are part of the shrinking result.

Shrinks towards zero(T). zero(T) produces no shrunk values.

PropCheck.shrink — Method

shrink(::T) where T <: Unsigned

Shrinks an unsigned value by masking out various part of its bitpattern.

Shrinks towards zero(T). zero(T) produces no shrunk values.

PropCheck.shrink — Method

shrink(::Vector{T}) where T

Shrinks a vector by either dropping an element, or shrinking an element. Does not modify its input.

Shrinks towards the empty vector, T[]. The empty vector produces no shrunk values.

PropCheck.shrinkTowards — Function

shrinkTowards(to::T) -> (x::T -> T[...])

Constructs a shrinker function that shrinks given values towards to.

PropCheck.str — Function

str(len[, alphabet])

Generates a string using the given genLen as a generator for the length. The default alphabet is typemin(Char):""[1], which is all representable Char values.

PropCheck.tuple — Method

tuple(genLen, genA)

Generates a tuple of a generated length, using the elements produced by genA.

PropCheck.unfold — Method

unfold(f, t)

Unfolds t into a Tree, by applying f to each root to create subtrees.

PropCheck.ExtentIntegrated — Type

ExtentIntegrated{T} <: AbstractIntegrated{T}

An integrated shrinker which has a bounds. The bounds can be accessed with the extent function.

PropCheck.Flatten — Type

Flatten with inferrable eltype. Requires the given iterators to have the same eltype.

PropCheck.IntegratedConst — Type

IntegratedConst{T,R,G} <: ExtentIntegrated{T}

An integrated shrinker describing a constant. The shrinker will always produce that value, which doesn't shrink.

PropCheck.IntegratedRange — Type

IntegratedRange{T,R,G,F} <: ExtentIntegrated{T}

An integrated shrinker describing a range of values.

The values created by this shrinker shrink according to the given shrinking function. The shrinking function must ensure that the produce values are always contained within the bounds of the given range.

PropCheck.IntegratedUnique — Type

IntegratedUnique(vec::Vector{T}, shrink::S) where {T,S}

An integrated shrinker, taking a vector vec. The shrinker will produce all unique values of vec in a random order before producing a value it returned before. The values produced by this shrinker shrink according to shrink.

PropCheck.IntegratedVal — Type

IntegratedVal(val::V, shrink::S) where {V,S}

An integrated shrinker, taking a value val. The shrinker will always produce val, which shrinks according to shrink.

If V <: Number, shrinking functions given to this must be producing values in

[typemin(V), v] if v > zero(V)

[v, typemax(V)] if v < zero(V)

no values if iszero(v)

This shrinker supports extent out of the box if V <: Number.

PropCheck.UniqueIterator — Type

UniqueIterator(itr, by=identity)

A lazy iterator over unique elements that have not been produced before.